VOXELUTIL: unified transformations

+++ src/modules/voxelutil/VolumeRotator.h
@@ -16,7 +16,7 @@ class Voxel;
 /**
  * @brief Rotate the given volume by the given angles in degree
  */
-extern RawVolume* rotateVolume(const RawVolume* source, const glm::vec3& angles, const glm::ivec3& pivot, bool increaseSize = true);
+extern RawVolume* rotateVolume(const RawVolume* source, const glm::vec3& angles, const glm::vec3& pivot, bool increaseSize = true);
 /**
  * @brief Rotate the given volume on the given axis by 90 degree. This method does not lose any voxels
  * @note The volume size might differ
